# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ugbanawaji API Docs",
        "description": "Ugbanawaji Application Programme Interface(Play Ground)",
        "termsOfService": "This API is subject to terms of services and maybe be changed without service notice.",
        "contact": {
            "name": "Ekenekiso Leonard Ugbanawaji",
            "url": "https://ugbanawaji.com",
            "email": "e.leonard@pensuh.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://www.pensuh.com/public/web/images/logo.png"
        }
    },
    "servers": [
        {
            "url": "https://api.ugbanawaji.com/api/v1",
            "description": "Ugbanawaji API Live Server"
        },
        {
            "url": "http://backend.test/api/v1",
            "description": "Ugbanawaji API Local Server. This server is attached to the local environment."
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Ugbanawaji API Local Server. This server is attached to the local environment."
        },
        {
            "url": "http://backend.test/api/v1"
        }
    ],
    "paths": {
...
        "/admin/employee/{employeeId}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "The employee resource",
                "description": "The employee resource.",
                "operationId": "Api/Admin/EmployeeController::show",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "The employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "The delete employee resource",
                "description": "The employee resource.",
                "operationId": "Api/Admin/EmployeeController::destroy",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "The employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/roles/{role}": {
            "get": {
                "tags": [
                    "Admin Roles"
                ],
                "summary": "The role resource",
                "description": "The role resource.",
                "operationId": "Api/Admin/RoleController::show",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Roles"
                ],
                "summary": "The role resource",
                "description": "The role resource.",
                "operationId": "Api/Admin/RoleController::destroy",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Roles"
                ],
                "summary": "Store Role resource.",
                "description": "Store Role resource.",
                "operationId": "Api/Admin/RoleController::update",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The Role resource id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create new Role",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            },
                            "example": {
                                "name": "The name",
                                "description": "The name"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true,
                                    "data": {
                                        "name": "The program resource",
                                        "created_at": "2022-09-08T12:29:54.000000Z",
                                        "updated_at": "2022-09-08T12:29:54.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/trilio/projects/{projectUuid}/activities": {
            "get": {
                "tags": [
                    "Activities",
                    "Trilio"
                ],
                "summary": "The resource collection",
                "description": "The resource collection",
                "operationId": "Trilio/Api/ActivityController::index",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search the resource by name or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "projectUuid",
                        "in": "query",
                        "description": "Scope the activities to project uuid. pass false if not required",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "The quantity",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActivityResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities",
                    "Trilio"
                ],
                "summary": "Create new activity",
                "description": "Create new activity",
                "operationId": "Trilio/Api/ActivityController::store",
                "parameters": [
                    {
                        "name": "projectUuid",
                        "in": "path",
                        "description": "The project uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create new activity",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityRequest"
                            },
                            "example": {
                                "name": "The brown fox",
                                "start_date": "2025-06-11 10:25:46",
                                "end_date": "2025-06-11 10:25:46",
                                "description": "The quick brown fox project test creation"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success.",
                                    "status": true,
                                    "data": {
                                        "name": "The brown fox",
                                        "uuid": "ashjvdHJSVDXSjdhHGVDGHJVGHD-AJBHGD",
                                        "id": "2",
                                        "status": "PENDING",
                                        "start_date": "2022-09-08T12:29:54.000000Z",
                                        "end_date": "2022-09-08T12:29:54.000000Z",
                                        "description": "The quick brown fox project test creation",
                                        "createdAt": "2022-09-08T12:29:54.000000Z",
                                        "updatedAt": "2022-09-08T12:29:54.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/trilio/projects/activities/{activityid}": {
            "get": {
                "tags": [
                    "Activities",
                    "Trilio"
                ],
                "summary": "The activity resource",
                "description": "The activity resource.",
                "operationId": "Trilio/Api/ActivityController::show",
                "parameters": [
                    {
                        "name": "Activityuid",
                        "in": "path",
                        "description": "The activity uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "success",
                                    "status": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleRequest": {
                "properties": {
                    "name": {
                        "description": "The role name.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ActivityResource": {
                "properties": {
                    "id": {
                        "description": "The activity Id.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The role name.",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "The activity uuid.",
                        "type": "string"
                    },
                    "projectId": {
                        "description": "The activity projectId.",
                        "type": "string"
                    },
                    "slug": {
                        "description": "The activity slug to support S.E.O.",
                        "type": "string"
                    },
                    "description": {
                        "description": "The activity description.",
                        "type": "string"
                    },
                    "start_date": {
                        "description": "The role start_date.",
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "description": "The activity end_date.",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "400": {
                "description": "Bad Request.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Bad Request.",
                            "errors": []
                        }
                    }
                }
            },
...
            "404": {
                "description": "Resource Not Found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Resource Not Found.",
                            "errors": []
                        }
                    }
                }
            },
...
            "422": {
                "description": "Unprocessable Entity.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Unprocessable Entity.",
                            "errors": []
                        }
                    }
                }
            },
...
            "500": {
                "description": "Server Error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Server Error.",
                            "errors": []
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backend.test/oauth/authorize",
                        "tokenUrl": "http://backend.test/api/v1/login",
                        "refreshUrl": "http://backend.test/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}